+++ oap-server/ai-pipeline/src/main/java/org/apache/skywalking/oap/server/ai/pipeline/services/api/HttpUriRecognition.java
@@ -52,6 +52,5 @@ public interface HttpUriRecognition extends Service {
     @Getter
     class HTTPUri {
         private final String name;
-        private final long matchedCounter;
     }
 }